move all sorts to sorts.h
