+++ lib/sorts.h
@@ -30,3 +30,24 @@ void SelectionSort(int unsorted[], size_t size){
 
     }
 }
+
+
+void InsertionSort(int unsorted[], size_t size){
+    for(int i = 1; i < size; i++){
+        int indexSwap = i;
+        for(int j = i-1; j >=0; j--){
+            if(unsorted[i] < unsorted[j]){
+                indexSwap = j;
+            }
+        }
+
+        int ToSwap = unsorted[i];
+
+        for(int k = i; k > indexSwap; k--){
+            unsorted[k] = unsorted[k-1];
+        }
+
+        unsorted[indexSwap] = ToSwap;
+
+    }
+}
